Relocate Season to model to Core
