+++ BreakingBadFan App/BreakingBadFan App/Home Screen/HomeViewController.swift
@@ -9,7 +9,8 @@ class HomeViewController: MainViewController {
     @IBOutlet private weak var quotesButton: UIButton!
     @IBOutlet private weak var logoutButton: UIButton!
     
-    private let apiManager = APIManager()
+    private let apiManager = Core.ApiManager
+    private let model = Core.Seasons
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -17,11 +18,12 @@ class HomeViewController: MainViewController {
         guard let loggedInAccount = AccountManager.loggedInAccount else { return }
         usernameLabel.text = loggedInAccount.username
         quotesButton.isEnabled = !loggedInAccount.favouriteQuotes.isEmpty
+        
     }
     
     @IBAction private func episodesButtonTapped(_ sender: UIButton) {
         print("🟢 episodesButtonDidTap")
-        proceedEpisodesScene()
+        fetchEpisodesToModel()
     }
     
     @IBAction private func charactersButtonTapped(_ sender: UIButton) {
@@ -49,5 +51,21 @@ class HomeViewController: MainViewController {
         dismiss(animated: true)
     }
     
+    private func fetchEpisodesToModel() {
+        apiManager.getEpisodes { result in
+            switch result {
+            case .success(let episodes):
+                print("🟢  Episodes didFetch")
+                self.model.episodes = episodes
+                
+                print("🟢🟢🟢 Season numbers \(self.model.seasonNumbers)")
+                
+                self.proceedEpisodesScene()
+                
+            case .failure(let error):
+                print("🔴 \(error)")
+            }
+        }
+    }
     
 }
