+++ BreakingBadFan App/BreakingBadFan App/Managers/Core.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct Core {
+    static let Seasons = SeasonModel()
+    static let ApiManager = APIManager()
+}
